requirements.txt
dj-static==0.0.6
Django==2.2.6
django-stdimage==5.0.1
gunicorn==19.9.0
Pillow==6.2.1
progressbar2==3.47.0
psycopg2-binary==2.8.4
python-utils==2.3.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
